refactor: Move from infrastructure/cloud to infrastructure/firebase ; Add firebase auth
