--- backend/src/infrastructure/cloud/firebase.ts
@@ -1,15 +1,0 @@
-import { FirebaseOptions, initializeApp } from "firebase/app";
-import dotenv from "dotenv";
-
-dotenv.config();
-
-const firebaseConfig: FirebaseOptions = {
-  apiKey: process.env.FIREBASE_API_KEY,
-  authDomain: process.env.FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.FIREBASE_PROJECT_ID,
-  storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.FIREBASE_MESSAGING_SENDER_ID,
-  appId: process.env.FIREBASE_APP_ID,
-};
-
-export const app = initializeApp(firebaseConfig);
